PDT-21f: User authorization complete
