+++ routes/userRoutes/users.py
@@ -1,14 +1,17 @@
-from routes.userRoutes.authorization import requires_auth
+from routes.authorization import requires_auth, AuthError, handle_auth_error
 from flask_cors import cross_origin
 from extensions import *
 
 userRoutes = Blueprint('userRoutes', __name__)
 
+# Error handler
+userRoutes.register_error_handler(AuthError, handle_auth_error)
+
 
 # Create a new user
 @userRoutes.route('/user', methods=['POST'])
 @cross_origin(headers=["Content-Type", "Authorization"])
-@requires_auth
+@requires_auth(audience="http://127.0.0.1:5000/user")
 def _create_user():
     user_collection = mongo_client.db.Users
     email = request.form.get('email')
@@ -21,7 +24,7 @@ def _create_user():
 # Get a user with the given ID
 @userRoutes.route('/user/<user_id>', methods=['GET'])
 @cross_origin(headers=["Content-Type", "Authorization"])
-@requires_auth
+@requires_auth(audience="http://127.0.0.1:5000/user")
 def _get_user(user_id):
     user_collection = mongo_client.db.Users
     user = user_collection.find_one({'_id': ObjectId(user_id)})
@@ -32,7 +35,7 @@ def _get_user(user_id):
 # Update the email address of a user with the given ID
 @userRoutes.route('/user/<user_id>', methods=['PUT'])
 @cross_origin(headers=["Content-Type", "Authorization"])
-@requires_auth
+@requires_auth(audience="http://127.0.0.1:5000/user")
 def _edit_user(user_id):
     user_collection = mongo_client.db.Users
     new_email = request.form.get('email')
@@ -44,7 +47,7 @@ def _edit_user(user_id):
 # Delete the user with the given ID
 @userRoutes.route('/user/<user_id>', methods=['DELETE'])
 @cross_origin(headers=["Content-Type", "Authorization"])
-@requires_auth
+@requires_auth(audience="http://127.0.0.1:5000/user")
 def _delete_user(user_id):
     user_collection = mongo_client.db.Users
     user = user_collection.delete_one({'_id': ObjectId(user_id)})
